requirements.txt
google-adk>=1.6.1
google-generativeai>=0.8.0
uiautomator2>=3.2.0
adb-shell>=0.4.0
python-dotenv>=1.0.0
Pillow>=10.0.0
google-cloud-core>=2.4.0
requests>=2.31.0 